// src/app/api/generated/model/update-consultorio-request.ts
/**
 * AKINE API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Cambios a aplicar sobre una sede. Los campos omitidos no se tocan
 */
export interface UpdateConsultorioRequest { 
    /**
     * Direccion. Cadena vacia la borra
     */
    addressLine?: string;
    /**
     * Email de contacto. Cadena vacia lo borra
     */
    contactEmail?: string;
    /**
     * Razon social. Cadena vacia la borra
     */
    legalName?: string;
    /**
     * Nuevo nombre. Unico entre las sedes vigentes del tenant
     */
    name?: string;
    /**
     * Telefono. Cadena vacia lo borra
     */
    phone?: string;
    /**
     * Nuevo intervalo por defecto de la agenda, en minutos
     */
    slotMinutes?: number;
    /**
     * Identificacion fiscal. Cadena vacia la borra
     */
    taxId?: string;
    /**
     * Nueva zona horaria IANA de la sede
     */
    timezone?: string;
    /**
     * Version leida del recurso. Si quedo vieja, la edicion se rechaza con 409 en vez de pisar el cambio de otro
     */
    version: number;
}